fix: resolve mypy type errors for Python 3.10 Linux CI
- Fix Mailgun config optional type checks
- Cast float to int for wait_time assignment
- Fix DataFrame column indexing with explicit list conversion
- Add Google config None check before accessing attributes
- Handle union types in pretalx utils with proper type guards
- Import and use QueryParams type in integration tests

These changes ensure mypy type checking passes on Python 3.10 Linux
where linting is configured to run in CI.

+++ src/pytanis/communication/mailgun_adapter.py
@@ -107,7 +107,7 @@ class MailgunAdapter(BaseMailClient):
             original_batch_size = self._client.batch_size
             original_wait_time = self._client.wait_time
             self._client.batch_size = 1  # Send one at a time
-            self._client.wait_time = 1.0 / rate_limit
+            self._client.wait_time = int(1.0 / rate_limit)
 
         try:
             for i, message in enumerate(messages):
